[metasrv] refactor: move test codes, nothing changed
